% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funloop.R
\name{superdooperfunlooper}
\alias{superdooperfunlooper}
\title{The Super Looper}
\usage{
superdooperfunlooper(funs = NULL)
}
\arguments{
\item{funs}{How many funs you're going to have.}
}
\value{
A vector of strings of funs.
}
\description{
The Super Looper
}
